// Coink/src/app/models/Goal.ts
/**
 * User goal
 */
export interface Goal {
  category_description: string;
  category_hash: string;
  category_id: number;
  current_balanse: number;
  daily_fill: number;
  goal_created_date: string;
  goal_description: string;
  goal_id: string;
  goal_imagen_cut: string;
  goal_limit_date: string;
  goal_missing_days: number;
  goal_status_id: number;
  goal_subcategory: string;
  goal_subcategory_id: number;
  goal_type: string;
  goal_value: number;
  saved_money: number;
  status: number;
  time_elapsed: number;
}