// Coink/src/app/models/Kid.ts
/**
 * User kid
 */
export interface Kid {
  authorization_id: number;
  balance: number;
  balance_total: number;
  count_goals: number;
  first_color: string;
  first_pocket: boolean;
  general_description: string;
  goal_balance: number;
  id_aliado: string;
  internal_description: string;
  logo: string;
  main_features: string;
  market_place: boolean;
  name_vault: string;
  nit_vault: string;
  restrictive: false;
  second_color: string;
  type_vault: number;
  update_term_conditions: boolean;
  vault_id: string;
}